requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-py==0.43.2
annotated-types==0.7.0
attrs==25.4.0
build==1.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
databento==0.72.0
databento-dbn==0.50.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
git-filter-repo==2.47.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.0
propcache==0.4.1
pyarrow==23.0.1
pybind11==3.0.2
pydantic==2.12.5
pydantic_core==2.41.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
setuptools==78.1.0
six==1.17.0
sseclient-py==1.9.0
sympy==1.14.0
TA-Lib==0.6.8
torch==2.12.0.dev20260301+rocm7.1
torchaudio==2.11.0.dev20260302+rocm7.1
torchvision==0.26.0.dev20260302+rocm7.1
triton-rocm==3.6.0+git9844da95
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yarl==1.23.0
zstandard==0.25.0
